Move useSnackBar to snackBar folder
